+++ conda_libmamba_solver/index.py
@@ -112,8 +112,9 @@ from .mamba_utils import logger_callback
 
 if TYPE_CHECKING:
     from collections.abc import Iterable
-    from typing import Literal
+    from typing import Any, Literal
 
+    from conda.common.path import PathsType
     from conda.gateways.repodata import RepodataState
     from libmambapy import QueryResult
     from libmambapy.solver.libsolv import RepoInfo
@@ -132,7 +133,7 @@ class _ChannelRepoInfo:
     url_no_cred: str
 
     @property
-    def canonical_name(self):
+    def canonical_name(self) -> str:
         if self.channel:
             return self.channel.canonical_name
         url_parts = self.url_no_cred.split("/")
@@ -158,7 +159,7 @@ class LibMambaIndexHelper:
         subdirs: Iterable[str] = (),
         repodata_fn: str = REPODATA_FN,
         installed_records: Iterable[PackageRecord] = (),
-        pkgs_dirs: Iterable[os.PathLike] = (),
+        pkgs_dirs: PathsType = (),
     ):
         platform_less_channels = []
         for channel in channels:
@@ -185,7 +186,7 @@ class LibMambaIndexHelper:
         self._set_repo_priorities()
 
     @classmethod
-    def from_platform_aware_channel(cls, channel: Channel):
+    def from_platform_aware_channel(cls, channel: Channel) -> LibMambaIndexHelper:
         if not channel.platform:
             raise ValueError(f"Channel {channel} must define 'platform' attribute.")
         subdir = channel.platform
@@ -208,7 +209,7 @@ class LibMambaIndexHelper:
                 count += len(self.db.packages_in_repo(repo))
         return count
 
-    def reload_channel(self, channel: Channel):
+    def reload_channel(self, channel: Channel) -> None:
         urls = {}
         for url in channel.urls(with_credentials=False, subdirs=self.subdirs):
             for repo_info in self.repos:
@@ -231,7 +232,7 @@ class LibMambaIndexHelper:
         }
         custom_channels = {
             name: LibmambaChannel(
-                url=CondaURL.parse(channel.base_url),
+                url=CondaURL.parse(channel.base_url.replace(" ", "%20")),
                 display_name=name,
                 platforms=set(self.subdirs),
             )
@@ -243,7 +244,7 @@ class LibMambaIndexHelper:
                 custom_channels.get(
                     channel.name,
                     LibmambaChannel(
-                        url=CondaURL.parse(channel.base_url),
+                        url=CondaURL.parse(channel.base_url.replace(" ", "%20")),
                         display_name=channel.name,
                         platforms=set(self.subdirs),
                     ),
@@ -272,6 +273,15 @@ class LibMambaIndexHelper:
     ) -> list[_ChannelRepoInfo]:
         if urls_to_channel is None:
             urls_to_channel = self._channel_urls()
+
+        # conda.common.url.path_to_url does not %-encode spaces
+        encoded_urls_to_channel = {}
+        for url, channel in urls_to_channel.items():
+            if url.startswith("file://"):
+                url = url.replace(" ", "%20")
+            encoded_urls_to_channel[url] = channel
+        urls_to_channel = encoded_urls_to_channel
+
         urls_to_json_path_and_state = self._fetch_repodata_jsons(tuple(urls_to_channel.keys()))
         channel_repo_infos = []
         for url_w_cred, (json_path, state) in urls_to_json_path_and_state.items():
@@ -310,7 +320,7 @@ class LibMambaIndexHelper:
             if seen_noauth.issuperset(noauth_urls):
                 continue
             auth_urls = [
-                url
+                url.replace(" ", "%20")
                 for url in channel.urls(with_credentials=True)
                 if url.endswith(tuple(self.subdirs))
             ]
@@ -440,7 +450,7 @@ class LibMambaIndexHelper:
                 log.debug("Ignored SOLV writing error for %s", channel_id, exc_info=exc)
         return repo
 
-    def _load_installed(self, records: Iterable[PackageRecord]) -> RepoInfo:
+    def _load_installed(self, records: Iterable[PackageRecord]) -> _ChannelRepoInfo:
         packages = [self._package_info_from_package_record(record) for record in records]
         repo = self.db.add_repo_from_packages(
             packages=packages,
@@ -452,7 +462,7 @@ class LibMambaIndexHelper:
             channel=None, repo=repo, url_w_cred="installed", url_no_cred="installed"
         )
 
-    def _load_pkgs_cache(self, pkgs_dirs: Iterable[os.PathLike]) -> list[RepoInfo]:
+    def _load_pkgs_cache(self, pkgs_dirs: PathsType) -> list[RepoInfo]:
         repos = []
         for path in pkgs_dirs:
             package_cache_data = PackageCacheData(path)
@@ -462,7 +472,8 @@ class LibMambaIndexHelper:
                 for record in package_cache_data.values()
             ]
             repo = self.db.add_repo_from_packages(packages=packages, name=path)
-            path_as_url = path_to_url(path)
+            # path_to_url does not %-encode spaces
+            path_as_url = path_to_url(path).replace(" ", "%20")
             repos.append(
                 _ChannelRepoInfo(
                     channel=None, repo=repo, url_w_cred=path_as_url, url_no_cred=path_as_url
@@ -506,7 +517,7 @@ class LibMambaIndexHelper:
             **extra,
         )
 
-    def _set_repo_priorities(self):
+    def _set_repo_priorities(self) -> None:
         has_priority = context.channel_priority in (
             ChannelPriority.STRICT,
             ChannelPriority.FLEXIBLE,
@@ -551,7 +562,7 @@ class LibMambaIndexHelper:
         self,
         queries: Iterable[str | MatchSpec] | str | MatchSpec,
         return_type: Literal["records", "dict", "raw"] = "records",
-    ) -> list[PackageRecord] | dict | QueryResult:
+    ) -> list[PackageRecord] | dict[str, Any] | QueryResult:
         if isinstance(queries, (str, MatchSpec)):
             queries = [queries]
         queries = list(map(str, queries))
@@ -563,7 +574,7 @@ class LibMambaIndexHelper:
         query: str | MatchSpec,
         tree: bool = False,
         return_type: Literal["records", "dict", "raw"] = "records",
-    ) -> list[PackageRecord] | dict | QueryResult:
+    ) -> list[PackageRecord] | dict[str, Any] | QueryResult:
         query = str(query)
         result = Query.depends(self.db, query, tree)
         return self._process_query_result(result, return_type)
@@ -573,7 +584,7 @@ class LibMambaIndexHelper:
         query: str | MatchSpec,
         tree: bool = False,
         return_type: Literal["records", "dict", "raw"] = "records",
-    ) -> list[PackageRecord] | dict | QueryResult:
+    ) -> list[PackageRecord] | dict[str, Any] | QueryResult:
         query = str(query)
         result = Query.whoneeds(self.db, query, tree)
         return self._process_query_result(result, return_type)
@@ -582,7 +593,7 @@ class LibMambaIndexHelper:
         """
         Returns all the package names that (might) depend on the passed specs
         """
-        explicit_pool = set()
+        explicit_pool: set[str] = set()
         for spec in specs:
             pkg_records = self.depends(spec.dist_str())
             for record in pkg_records:
@@ -593,7 +604,7 @@ class LibMambaIndexHelper:
         self,
         result: QueryResult,
         return_type: Literal["records", "dict", "raw"] = "records",
-    ) -> list[PackageRecord] | dict | QueryResult:
+    ) -> list[PackageRecord] | dict[str, Any] | QueryResult:
         if return_type == "raw":
             return result
         result = result.to_dict()